Moved database logic to separate module.
